refactor(folders): restructure hooks folder
